requirements.txt
cohere 
gradio
python-dotenv